% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefixEncoding.R
\name{aggregation_encoding}
\alias{aggregation_encoding}
\title{Aggregation Encoding}
\usage{
aggregation_encoding(prefix_eventLog)
}
\arguments{
\item{eventLog}{A event log with prefix.}
}
\value{
event log with encoding.
}
\description{
Aggregation encoding. In contrast to the last-state encoding, aggregation encoding considers all events in the prefix of a case, rather than considering only the last m events. Each attribute is encoded into one or multiple features using different aggregation functions depending on the datatype of the attribute.
If an attribute is of numerical type then we map this attribute into one feature by means of a numerical aggregation function such as sum, average, minimum, or maximum.
f an attribute is categorical we map this attribute into one feature for every value in the attribute’s domain by applying the “count” aggregation function.
}
\examples{
library(bupaR)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
enrichEventlogEncoding <- aggregation_encoding(prefix_eventLog = eventdata,Window=3)
}
