% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{BuildModel}
\alias{BuildModel}
\title{Build Model with Encoding Event Log Data}
\usage{
BuildModel(
  TheModel = rand_forest,
  engine = "ranger",
  PrefixData,
  predictmode = "regression"
)
}
\arguments{
\item{TheModel}{Model function in Tidymodels,such as boost_tree,rand_forest.}

\item{engine}{engine of model.}

\item{PrefixData}{Encoding event log data.}

\item{predictmode}{Type of model,regression or classification.}
}
\value{
Return A list about train dataset,test dataset and model result.
}
\description{
Build all kind of regression and classification model.This function is base on Tidymodel package.
}
\examples{
library(bupar)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
enrichEventlogEncoding <- lastStateEncoding(prefix_eventLog = eventdata)
enrichEventlogEncoding.1 <- enrichEventlogEncoding \%>\% as.data.frame()
enrichEventlogEncoding.1 <- enrichEventlogEncoding.1 \%>\% select(-patient,-handling_id,-.order,-registration_type)
library(lubridate)
enrichEventlogEncoding.1 <- enrichEventlogEncoding.1 \%>\% mutate(year=year(time),month=month(time),day = day(time),week = week(time),hour = hour(time))
enrichEventlogEncoding.1 <- enrichEventlogEncoding.1 \%>\% select(-time)
enrichEventlogEncoding.1 <- enrichEventlogEncoding.1 \%>\% filter(!is.na(predictor))
enrichEventlogEncoding.1$predictor <- enrichEventlogEncoding.1$predictor \%>\% as.character() \%>\% as.factor()
EventLogModel <- BuildModel(TheModel = rand_forest,engine = "ranger",PrefixData = enrichEventlogEncoding.1,predictmode = "classification")
pre <- predict(object = EventLogModel[[3]], EventLogModel[["test"]],type = "class")
}
