% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichEventlog.R
\name{createProcessIndicatorsWithPrefix}
\alias{createProcessIndicatorsWithPrefix}
\title{prefix and create indicators  for evnetlog}
\usage{
createProcessIndicatorsWithPrefix(eventLog, prefix_num = 5, mode = "activity")
}
\arguments{
\item{eventLog}{A event log which class is eventlog from bupaR}

\item{prefix_num}{prefix length of trace}

\item{mode}{create indicators,for now，have two option :activity or duration}
}
\value{
event log with prefix and indicators
}
\description{
this function transform original event log to prefix event log and create indicators for predicting
}
\examples{
library(bupar)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
}
