% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichEventlog.R
\name{eventAnalyse}
\alias{eventAnalyse}
\title{Analyse event in control-flow perspective, time perspective and resource perspective}
\usage{
eventAnalyse(eventlog, perspective = c("controlflow", "time", "resource"))
}
\arguments{
\item{eventlog}{A event log}

\item{perspective}{Analysis perspective}
}
\value{
A list of result
}
\description{
Analyse event in control-flow perspective, time perspective and resource perspective
}
\examples{
x <- eventAnalyse(eventlog = patients,perspective = "time")
}
