% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceBucketing.R
\name{hierarchicalClusteringBucketing}
\alias{hierarchicalClusteringBucketing}
\title{Hierarchical Cluster Bucketing}
\usage{
hierarchicalClusteringBucketing(
  prefix_eventLog,
  dist_methods = "lv",
  cluster_num = 3,
  cluster_methods = "agnes",
  ...
)
}
\arguments{
\item{prefix_eventLog}{A event log with prefix.}

\item{dist_methods}{Measure the distance of trace.Option inclcude:"osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",}

\item{cluster_num}{Cluster number.}

\item{cluster_methods}{Cluster methods.Option include:agnes,diana.}
}
\value{
event log with Cluster label
}
\description{
Cluster trace by hierarchical clustering.
}
\examples{
library(bupaR)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
enrichEventlogEncoding <- hierarchicalClusteringBucketing(prefix_eventLog = eventdata)
}
