% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceBucketing.R
\name{KMmeanclusterBucketing}
\alias{KMmeanclusterBucketing}
\title{Kmeans Cluster Bucketing}
\usage{
KMmeanclusterBucketing(
  prefix_eventLog,
  dist_methods = "lv",
  cluster_methods = "clara",
  cluster_num = 3,
  ...
)
}
\arguments{
\item{prefix_eventLog}{A event log with prefix.}

\item{dist_methods}{Measure the distance of trace.Option inclcude:"osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",}

\item{cluster_methods}{Cluster methods.Option include:clara,fanny,pam.}

\item{cluster_num}{Cluster number.}
}
\value{
event log with Cluster label
}
\description{
Cluster trace by Kmeans Cluster.
}
\examples{
library(bupar)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
enrichEventlogEncoding <- KMmeanclusterBucketing(prefix_eventLog = eventdata)
}
