% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefixEncoding.R
\name{lastNStateIndexBasedEncoding}
\alias{lastNStateIndexBasedEncoding}
\title{Last N State Index Based Encoding}
\usage{
lastNStateIndexBasedEncoding(prefix_eventLog, Window = 3)
}
\arguments{
\item{Window}{The Window length of Trace.}

\item{eventLog}{A event log with prefix.}
}
\value{
event log with encoding.
}
\description{
The last M available snapshot of the data is used ,And The idea of indexbased encoding is to use all possible information (including the order) in the trace, generating one feature per each event attribute per each executed event (each index).
}
\examples{
library(bupaR)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
enrichEventlogEncoding <- lastNStateIndexBasedEncoding(prefix_eventLog = eventdata,Window=3)
}
