% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefixEncoding.R
\name{lastStateEncoding}
\alias{lastStateEncoding}
\title{Last State Encoding}
\usage{
lastStateEncoding(prefix_eventLog)
}
\arguments{
\item{eventLog}{A event log with prefix.}
}
\value{
event log with encoding.
}
\description{
In this encoding method, only the last available snapshot of the data is used.
}
\examples{
library(bupaR)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
enrichEventlogEncoding <- lastStateEncoding(prefix_eventLog = eventdata)
}
