% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceBucketing.R
\name{optimalNumberOfCluster}
\alias{optimalNumberOfCluster}
\title{Gap Statistic for Estimating the Number of Clusters}
\usage{
optimalNumberOfCluster(
  prefix_eventLog,
  cluster_methods = "fanny",
  dist_methods = "lv",
  max_cluster_num = 5,
  ...
)
}
\arguments{
\item{prefix_eventLog}{A event log with prefix.}

\item{cluster_methods}{Cluster methods.Option include:clara,fanny,pam.}

\item{dist_methods}{Measure the distance of trace.Option inclcude:"osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",}

\item{max_cluster_num}{Max cluster number.}
}
\value{
The number of cluster.
}
\description{
Use gap statistic to estimating the number of clusters
}
\examples{
library(bupar)
library(ppmr)
eventdata <- enrichEventlog(eventLog = patients,prefix_num = 4,mode = "activity")
clusternum <- optimalNumberOfCluster(prefix_eventLog = eventdata,max_cluster_num=5,B=100)
}
