\name{PMR-package}
\alias{PMR-package}
\alias{PMR}
\docType{package}
\title{\packageTitle{PMR}}
\description{PMR is an R package for efficient statitical inferecne of two-sample MR analysis. It can account for the correlatded instruments and the horizontal pleiotropy, and can provide the accurate estimates of both causal effect and horizontal pleitropy effect as well as the two corresponding p values.There are two main functions in PMR package, one is PMR_individual for individual level data, the other is PMR_summary for summary data}
\details{
\tabular{ll}{
Package: \tab PMR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2019-08-04\cr
License: \tab GPL-3\cr
}
}
\author{
   Zhongshang Yuan, Xiang Zhou
  
  Maintainer: Zhongshang Yuan <yuanzhongshang@sdu.edu.cn>
}

% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}