% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PMR_individual_rcpp}
\alias{PMR_individual_rcpp}
\title{PMR Individual Rcpp Helper}
\usage{
PMR_individual_rcpp(yin, zin, x1in, x2in, gammain, alphain, max_iterin,
  epsin)
}
\arguments{
\item{yin}{standardized exposure vector (e.g. gene expression in TWAS).}

\item{zin}{standardized complex trait vector.}

\item{x1in}{standardized cis-genotype matrix in eQTL data.}

\item{x2in}{standardized cis-genotype matrix in GWAS data.}

\item{gammain}{indicator variable for constrained PMR model, with 1 for the null 
model that there is no horiozntal pleiotropy.}

\item{alphain}{indicator variable for constrained PMR model, with 1 for the null 
model that there is no causal effect.}

\item{max_iterin}{The maximum iteration.}

\item{epsin}{The convergence tolerance of the absolute value of the difference 
between the nth and (n+1)th loglikelihood.}
}
\value{
a list of estimates of model parameters, including the causal effect 
\code{alpha}, the horizontal pleiotropy effect \code{gamma}, and the two 
corresponding p values
}
\description{
Individual helper function which does the work in C++, returns it to R. Call
PMR_individual to use the function.
}
