% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PMR_summary_rcpp}
\alias{PMR_summary_rcpp}
\title{PMR Summary Rcpp Helper}
\usage{
PMR_summary_rcpp(betaxin, betayin, Sigma1sin, Sigma2sin, samplen1,
  samplen2, gammain, alphain, max_iterin, epsin)
}
\arguments{
\item{betaxin}{the cis-SNP effect size vector for one specific gene in eQTL 
data, which must be calculated based on both the standardized gene expression 
value and the standardized cis-genotype matrix.}

\item{betayin}{the cis-SNP effect size vector for one specific gene in GWAS 
data, which be calculated based on both the standardized complex trait value 
and the standardized cis-genotype matrix.}

\item{Sigma1sin}{the LD matrix in eQTL data.}

\item{Sigma2sin}{the LD matrix in GWAS data.Both \code{Sigma2sin} and 
\code{sigma1sin} are often the same from the reference panel.}

\item{samplen1}{the sample size of eQTL data.}

\item{samplen2}{the sample size of GWAS data.}

\item{gammain}{indicator variable for constrained model, with 1 for the null 
model that there is no horiozntal pleiotropy.}

\item{alphain}{indicator variable for constrained model, with 1 for the null 
model that there is no causal effect.}

\item{max_iterin}{The maximum iteration.}

\item{epsin}{The convergence tolerance of the absolute value of the difference 
between the nth and (n+1)th loglikelihood.}
}
\value{
A list of estimates of model parameters, including the causal effect 
\code{alpha}, the horizontal pleiotropy effect \code{gamma}, and the two corresponding p values.
}
\description{
Summary helper function which does the work in C++, returns it to R. Call
PMR_summary to use the function.
}
