% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichEventlog.R
\name{createProcessIndicatorsWithPrefix}
\alias{createProcessIndicatorsWithPrefix}
\title{Prefix And Create Indicators  For Evnetlog}
\usage{
createProcessIndicatorsWithPrefix(eventLog, prefix_num = 5, mode = "activity")
}
\arguments{
\item{eventLog}{A event log which class is eventlog from bupaR.}

\item{prefix_num}{Prefix length of trace.}

\item{mode}{Create indicators,for now，have two option :activity or duration.}
}
\value{
Event log with prefix and indicators.
}
\description{
This function transform original event log to prefix event log and create indicators for predicting.
}
\examples{
library(bupar)
library(ppmr)
eventdata <- createProcessIndicatorsWithPrefix(eventLog = patients,prefix_num = 4,mode = "activity")
}
