% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Examplesummary}
\alias{Examplesummary}
\title{Summary level dataset}
\format{A list contains the following objects:
\describe{
  \item{betax}{the cis-SNP effect size vector for one specific gene in eQTL data.}
  \item{betay}{the cis-SNP effect size vector for one specific gene in GWAS data.}
  \item{Sigma1}{the LD matrix in eQTL data.}
  \item{Sigma2}{the LD matrix in GWAS data.}
  \item{n1}{the sample size of eQTL data.}
  \item{n2}{the sample size of GWAS data.}
}}
\usage{
Examplesummary
}
\description{
A simulated summary level dataset for PMR
}
\keyword{datasets}
